Work by Ryan Govostes and me which improves the behavior of the RateQueue.
The RateClassMonitor's getPossibleCmdCount() is initially consulted to determine how many SnacRequests we should attempt to send. (This returns 0 at a certain degree of rate limiting... but we treat that as a value of 1 to avoid an infinite loop as indicated in the comments.)

Then, as each SnacRequest is actually dispatched, updateRate() is called with the current time.  isReady() is checked within the loop such that if updateRate() means that we're no longer ready to send, we stop emptying out the queue.  Remaining SnacRequests are re-enqueued at the front of the queue for sending when we're again ready.

This fixes the rate limiting problems for me.  It probably needs refinement -- particularly in that getPossibleCmdCount()'s logic may not be perfect since theoretically its return value _should_ predict how many we can send before isReady() becomes false.