When a ClientFlapConn disconnects or fails, AsynchronousFlapProcessor.breakdown() is now called on its AsynchronousFlapProcessor, and flapProcessor instance variable is cleared.
AsynchronousFlapProcessor.breakdown() causes the weak reference its thread uses to determine when to die to be immediately cleared.  It also detaches its thread and clears its observers lists.

This is necessary because AsynchronousFlapProcessor creates a thread which continues until the AsynchronousFlapProcessor itself is GC'd.  As I understand it based on YourKit Java Profiler analysis, that never happened previously,  because the observers list for the AsynchronousFlapProcessor includes an OscarConnection which keeps track of the ClientFlapConn which keeps track of the AsynchronousFlapProcessor... a loop starting with a GC root.

The ClientFlapConn now breaks that vicious cycle manually, allowing the thread to die.

This does not solve every leak, but it appears to fix the biggest one we've observed in Adium 1.0b14.